Initial commit
Welcome to the first commit in my gem: capybara-screenshot-annotated.
This gem will make it easy for technical writers and developers to
quickly automate screenshots for end user manuals and support articles.

The initial functionality allows you to create a cursor (using
JavaScript) and quickly put it on a link on the page that features
specific text. For example:

```
describe "a search query on duckduckgo.com" do
  it "shows some good results for 'ethiopian coffee'" do
    visit '/?q=ethiopian+coffee'
    create_cursor "pointer"
    hover_on_link "Best Ethiopian coffee varieties"
    page.save_screenshot
  end
end
```